Fixing transmission of 32bit values

+++ src/ECPacket.h
@@ -69,7 +69,7 @@ class CECTag {
 			// Retrieving special data types
 		uint8		GetInt8Data(void) const { return *((uint8 *)m_tagData); }
 		uint16		GetInt16Data(void) const { return ENDIAN_SWAP_16(*((uint16 *)m_tagData)); }
-		uint32		GetInt32Data(void) const { return ENDIAN_SWAP_32(*((uint16 *)m_tagData)); }
+		uint32		GetInt32Data(void) const { return ENDIAN_SWAP_32(*((uint32 *)m_tagData)); }
 		wxString	GetStringData(void) const { return wxString(wxConvUTF8.cMB2WC((const char *)m_tagData), aMuleConv); }
 		EC_IPv4_t *	GetIPv4Data(void) const;
 	protected:
